rest-api

+++ WebDev2-master/src/main/java/com/hannahjoy/car/api/CarRestController.java
@@ -4,7 +4,9 @@ import com.hannahjoy.car.DTO.CarDTO;
 import com.hannahjoy.car.Model.Car;
 import com.hannahjoy.car.Service.CarService;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -15,16 +17,19 @@ public class CarRestController {
     private final CarService carService;
 
     public CarRestController(CarService carService) {
+
         this.carService = carService;
     }
 
     @GetMapping("/cars")
     public List<Car> getCars() {
+
         return carService.getAllCars();
     }
 
     @GetMapping("/{id}")
     public Car getCarById(@PathVariable int id) {
+
         return carService.getCarById(id);
     }
 
@@ -36,13 +41,19 @@ public class CarRestController {
 
     @PutMapping("/cars/{id}")
     public Car updateCar(@PathVariable int id, @Valid @RequestBody CarDTO carDTO) {
-        carService.update(id, carDTO);
-        return carService.getCarById(id);
+        Car updateCar = carService.getCarById(id);
+        if(updateCar == null){
+            throw new ResponseStatusException ( HttpStatus.NOT_FOUND, "Car with ID " + id + " not found");
+        }
+        return carService.updateCar(id, carDTO);
     }
 
     @DeleteMapping("/cars/{id}")
     public Car deleteCar(@PathVariable int id) {
+        if(carService.getCarById(id) == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Car with ID " + id + " not found");
+        }
         carService.delete(id);
-        return carService.getCarById(id);
+        return null;
     }
 }
